+++ options/db_options.h
@@ -76,6 +76,7 @@ struct ImmutableDBOptions {
   bool dump_malloc_stats;
   bool avoid_flush_during_recovery;
   bool allow_ingest_behind;
+  bool preserve_deletes;
   bool concurrent_prepare;
   bool manual_wal_flush;
   bool seq_per_batch;